#18: moves some localization classes to localization package
