+++ iped-engine/src/main/java/dpf/sp/gpinf/indexer/config/RegexTaskConfig.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 import org.apache.commons.lang3.tuple.Pair;
 
-import dpf.sp.gpinf.indexer.Messages;
+import dpf.sp.gpinf.indexer.localization.Messages;
 import dpf.sp.gpinf.indexer.util.IPEDException;
 import dpf.sp.gpinf.indexer.util.Util;
